@@ -31,58 +31,53 @@
     "postversion": "git push && git push --tags && npm publish"
   },
   "dependencies": {
-    "@angular/animations": "^4.2.2",
-    "@angular/common": "^4.2.2",
-    "@angular/compiler": "^4.2.2",
-    "@angular/core": "^4.2.2",
-    "@angular/forms": "^4.2.2",
-    "@angular/http": "^4.2.2",
-    "@angular/material": "^2.0.0-beta.6",
-    "@angular/platform-browser": "^4.2.2",
-    "@ngx-translate/core": "^7.0.0",
-    "@ngx-translate/http-loader": "^0.1.0",
-    "buffer": "^5.0.6",
+    "@angular/animations": "^4.3.0",
+    "@angular/cdk": "^2.0.0-beta.8",
+    "@angular/common": "^4.3.0",
+    "@angular/compiler": "^4.3.0",
+    "@angular/core": "^4.3.0",
+    "@angular/forms": "^4.3.0",
+    "@angular/http": "^4.3.0",
+    "@angular/material": "^2.0.0-beta.8",
+    "@angular/platform-browser": "^4.3.0",
+    "@angular/platform-browser-dynamic": "^4.3.0",
+    "@angular/router": "^4.3.0",
+    "@ngx-translate/core": "^7.1.0",
+    "@ngx-translate/http-loader": "^1.0.1",
     "core-js": "^2.4.1",
-    "devextreme": "^17.1.3",
-    "devextreme-angular": "^17.1.3",
-    "hammerjs": "^2.0.8",
-    "igo2": "^0.7.0",
-    "rxjs": "^5.4.1",
-    "zone.js": "^0.8.12"
+    "devextreme": "^17.1.4",
+    "devextreme-angular": "^17.1.4",
+    "igo2": "^0.8.1",
+    "rxjs": "^5.4.2",
+    "zone.js": "^0.8.16"
   },
   "devDependencies": {
-    "@angular/cli": "^1.1.2",
-    "@angular/compiler-cli": "^4.2.2",
-    "@angular/platform-browser-dynamic": "^4.2.2",
-    "@angular/platform-server": "^4.2.2",
-    "@angular/router": "^4.2.2",
-    "@types/jasmine": "^2.5.52",
-    "@types/node": "^7.0.31",
-    "@types/openlayers": "^4.1.1",
+    "@angular/cli": "1.2.6",
+    "@angular/compiler-cli": "^4.3.0",
+    "@angular/language-service": "^4.3.0",
+    "@types/jasmine": "~2.5.53",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~6.0.60",
+    "@types/openlayers": "^4.1.4",
     "angular2-template-loader": "^0.6.2",
     "awesome-typescript-loader": "^3.1.3",
-    "codelyzer": "^3.0.1",
-    "css-loader": "^0.28.0",
+    "codelyzer": "~3.0.1",
     "gulp": "^3.9.1",
     "gulp-replace": "^0.5.4",
     "gulp-stylus": "^2.6.0",
-    "jasmine-core": "^2.6.3",
-    "jasmine-spec-reporter": "^4.1.0",
-    "karma": "^1.7.0",
-    "karma-chrome-launcher": "^2.1.1",
+    "jasmine-core": "~2.6.2",
+    "jasmine-spec-reporter": "~4.1.0",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.1.2",
-    "raw-loader": "^0.5.1",
-    "stylus": "^0.54.5",
-    "stylus-loader": "^3.0.1",
+    "protractor": "~5.1.2",
     "to-string-loader": "^1.1.5",
-    "ts-node": "^3.0.6",
-    "tslint": "^5.4.3",
-    "typescript": "^2.3.4",
-    "webpack": "^2.6.1",
-    "webpack-merge": "^4.1.0"
+    "ts-node": "~3.0.4",
+    "tslint": "~5.3.2",
+    "typescript": "~2.3.3",
+    "webpack": "^3.4.1"
   }
 }
